Finished Model class, User is now extended from Model with generic type of UserProps interface
