+++ src/models/Model.ts
@@ -1,4 +1,8 @@
-import { AxiosPromise } from 'axios';
+import { AxiosPromise, AxiosResponse } from 'axios';
+
+interface HasId {
+  id?: number
+}
 interface ModelAttributes<T> {
   get<K extends keyof T>(key: K): T[K]
   set(value: T): void
@@ -15,6 +19,49 @@ interface Events {
   trigger(eventName: string): void
 }
 
-export class Model {
+export class Model<T extends HasId> {
+  constructor(
+    private attributes: ModelAttributes<T>,
+    private events: Events,
+    private sync: Sync<T>
+  ) {}
 
+  get on() {
+    return this.events.on;
+  }
+
+  get trigger() {
+    return this.events.trigger
+  }
+
+  get get() {
+    return this.attributes.get;
+  }
+
+  set(update: T): void {
+    this.attributes.set(update)
+    this.events.trigger('change')
+  }
+
+  fetch(): void {
+    const id = this.get('id')
+
+    if (typeof id !== 'number') {
+      throw new Error("Cannout  fetch without and id")
+    }
+
+    this.sync.fetch(id).then((response: AxiosResponse): void => {
+      this.set(response.data)
+    })
+  }
+
+  save(): void {
+    this.sync.save(this.attributes.getAll())
+      .then((response: AxiosResponse): void => {
+        this.trigger('save')
+      })
+      .catch(() => {
+        this.trigger('error')
+      })
+  }
 }
