+++ src/models/User.ts
@@ -1,3 +1,4 @@
+import { Model } from './Model';
 import { Attributes } from './Attributes';
 import { Sync } from './Sync';
 import { Eventing } from './Eventing';
@@ -13,51 +14,4 @@ export interface UserProps {
 
 const rootUrl = 'http://localhost:3000/users'
 
-export class User {
-  public events: Eventing = new Eventing()
-  public sync: Sync<UserProps> = new Sync<UserProps>(rootUrl)
-  public attributes: Attributes<UserProps>
-
-  constructor(attrs: UserProps) {
-    this.attributes = new Attributes<UserProps>(attrs)
-  }
-
-  get on() {
-    return this.events.on;
-  }
-
-  get trigger() {
-    return this.events.trigger
-  }
-
-  get get() {
-    return this.attributes.get;
-  }
-
-  set(update: UserProps): void {
-    this.attributes.set(update)
-    this.events.trigger('change')
-  }
-
-  fetch(): void {
-    const id = this.get('id')
-
-    if (typeof id !== 'number') {
-      throw new Error("Cannout  fetch without and id")
-    }
-
-    this.sync.fetch(id).then((response: AxiosResponse): void => {
-      this.set(response.data)
-    })
-  }
-
-  save(): void {
-    this.sync.save(this.attributes.getAll())
-      .then((response: AxiosResponse): void => {
-        this.trigger('save')
-      })
-      .catch(() => {
-        this.trigger('error')
-      })
-  }
-}
+export class User extends Model<UserProps> {}
